+++ packages/lit-html/src/lit-html.ts
@@ -122,10 +122,15 @@ const HTML_RESULT = 1;
 const SVG_RESULT = 2;
 
 /** TemplatePart types */
-const ATTRIBUTE_PART = 1;
-const NODE_PART = 2;
-const ELEMENT_PART = 3;
-const COMMENT_PART = 4;
+// TODO (justinfagnani): since these are exported, consider shorter names,
+// like just `ATTRIBUTE`.
+export const ATTRIBUTE_PART = 1;
+export const NODE_PART = 2;
+export const PROPERTY_PART = 3;
+export const BOOLEAN_ATTRIBUTE_PART = 4;
+export const EVENT_PART = 5;
+const ELEMENT_PART = 6;
+const COMMENT_PART = 7;
 
 type ResultType = typeof HTML_RESULT | typeof SVG_RESULT;
 
@@ -187,13 +192,36 @@ export const nothing = {};
  */
 const templateCache = new Map<TemplateStringsArray, Template>();
 
-export type DirectiveClass = {new (part: Part): Directive};
+export type NodePartInfo = {
+  readonly type: typeof NODE_PART;
+};
+
+export type AttributePartInfo = {
+  readonly type:
+    | typeof ATTRIBUTE_PART
+    | typeof PROPERTY_PART
+    | typeof BOOLEAN_ATTRIBUTE_PART
+    | typeof EVENT_PART;
+  strings?: ReadonlyArray<string>;
+  name: string;
+  tagName: string;
+};
+
+/**
+ * Information about the part a directive is bound to.
+ *
+ * This is useful for checking that a directive is attached to a valid part,
+ * such as with directive that can only be used on attribute bindings.
+ */
+export type PartInfo = NodePartInfo | AttributePartInfo;
+
+export type DirectiveClass = {new (part: PartInfo): Directive};
 
 /**
  * This utility type extracts the signature of a directive class's render()
  * method so we can use it for the type of the generated directive function.
  */
-type DirectiveProps<C extends DirectiveClass> = Parameters<
+export type DirectiveParameters<C extends DirectiveClass> = Parameters<
   InstanceType<C>['render']
 >;
 
@@ -203,15 +231,18 @@ type DirectiveProps<C extends DirectiveClass> = Parameters<
  */
 type DirectiveResult<C extends DirectiveClass = DirectiveClass> = {
   _$litDirective$: C;
-  values: DirectiveProps<C>;
+  values: DirectiveParameters<C>;
 };
 
 /**
  * Creates a user-facing directive function from a Directive class. This
  * function has the same parameters as the directive's render() method.
+ *
+ * WARNING: The directive and part API changes are in progress and subject to
+ * change in future pre-releases.
  */
 export const directive = <C extends DirectiveClass>(c: C) => (
-  ...values: DirectiveProps<C>
+  ...values: DirectiveParameters<C>
 ): DirectiveResult<C> => ({
   _$litDirective$: c,
   values,
@@ -223,6 +254,10 @@ export interface RenderOptions {
    * useful to set this to the host component rendering a template.
    */
   readonly eventContext?: EventTarget;
+  /**
+   * A DOM node before which to render content in the container.
+   */
+  readonly renderBefore?: ChildNode | null;
 }
 
 /**
@@ -236,11 +271,15 @@ export const render = (
   container: HTMLElement | DocumentFragment,
   options?: RenderOptions
 ) => {
-  let part: NodePart = (container as any).$lit$;
+  const partOwnerNode = options?.renderBefore ?? container;
+  let part: NodePart = (partOwnerNode as any).$lit$;
   if (part === undefined) {
-    const marker = createMarker();
-    container.appendChild(marker);
-    (container as any).$lit$ = part = new NodePart(marker, null, options);
+    const endNode = options?.renderBefore ?? null;
+    (partOwnerNode as any).$lit$ = part = new NodePart(
+      container.insertBefore(createMarker(), endNode),
+      endNode,
+      options
+    );
   }
   part._setValue(value);
 };
@@ -254,6 +293,14 @@ const walker = d.createTreeWalker(d);
 // Interfaces and type aliases can be interleaved freely.
 //
 
+/**
+ * Base class for creating custom directives. Users should extend this class,
+ * implement `render` and/or `update`, and then pass their subclass to
+ * `directive`.
+ *
+ * WARNING: The directive and part API changes are in progress and subject to
+ * change in future pre-releases.
+ */
 export abstract class Directive {
   abstract render(...props: Array<unknown>): unknown;
   update(_part: Part, props: Array<unknown>): unknown {
@@ -530,9 +577,9 @@ class TemplateInstance {
         i++;
         continue;
       }
-      if ((part as AttributePart).__strings !== undefined) {
+      if ((part as AttributePart).strings !== undefined) {
         (part as AttributePart)._setValue(values, i);
-        i += (part as AttributePart).__strings!.length - 1;
+        i += (part as AttributePart).strings!.length - 1;
       } else {
         (part as NodePart)._setValue(values[i++]);
       }
@@ -581,6 +628,7 @@ export type Part =
   | BooleanAttributePart;
 
 export class NodePart {
+  readonly type = NODE_PART;
   _value: unknown;
   protected __directive?: Directive;
 
@@ -591,6 +639,9 @@ export class NodePart {
   ) {}
 
   _setValue(value: unknown): void {
+    // TODO (justinfagnani): when setting a non-directive over a directive,
+    // we don't yet clear this.__directive.
+    // See https://github.com/Polymer/lit-html/issues/1286
     if (isPrimitive(value)) {
       if (value !== this._value) {
         this.__commitText(value);
@@ -620,7 +671,7 @@ export class NodePart {
     const directive = value._$litDirective$;
     if (this.__directive?.constructor !== directive) {
       this.__clear();
-      this.__directive = new directive(this);
+      this.__directive = new directive(this as NodePartInfo);
     }
     // TODO (justinfagnani): To support nested directives, we'd need to
     // resolve the directive result's values. We may want to offer another
@@ -640,7 +691,7 @@ export class NodePart {
     // Make sure undefined and null render as an empty string
     // TODO: use `nothing` to clear the node?
     value ??= '';
-    // TODO(justinfagnani): Can we just check if this.__value is primitive?
+    // TODO(justinfagnani): Can we just check if this._value is primitive?
     if (
       node !== null &&
       node.nodeType === 3 /* Node.TEXT_NODE */ &&
@@ -743,6 +794,11 @@ export class NodePart {
 }
 
 export class AttributePart {
+  readonly type = ATTRIBUTE_PART as
+    | typeof ATTRIBUTE_PART
+    | typeof PROPERTY_PART
+    | typeof BOOLEAN_ATTRIBUTE_PART
+    | typeof EVENT_PART;
   readonly element: HTMLElement;
   readonly name: string;
 
@@ -751,9 +807,13 @@ export class AttributePart {
    * static strings of the interpolation. For single-value, complete bindings,
    * this is undefined.
    */
-  readonly __strings?: ReadonlyArray<string>;
-  __value: unknown | Array<unknown> = nothing;
+  readonly strings?: ReadonlyArray<string>;
+  _value: unknown | Array<unknown> = nothing;
   private __directives?: Array<Directive>;
+
+  get tagName() {
+    return this.element.tagName;
+  }
 
   constructor(
     element: HTMLElement,
@@ -764,10 +824,10 @@ export class AttributePart {
     this.element = element;
     this.name = name;
     if (strings.length > 2 || strings[0] !== '' || strings[1] !== '') {
-      this.__value = new Array(strings.length - 1).fill(nothing);
-      this.__strings = strings;
+      this._value = new Array(strings.length - 1).fill(nothing);
+      this.strings = strings;
     } else {
-      this.__value = nothing;
+      this._value = nothing;
     }
   }
 
@@ -786,7 +846,9 @@ export class AttributePart {
       // or check length.
       let directive: Directive = (this.__directives ??= [])[i];
       if (directive?.constructor !== directiveCtor) {
-        directive = this.__directives[i] = new directiveCtor(this);
+        directive = this.__directives[i] = new directiveCtor(
+          this as AttributePartInfo
+        );
       }
       // TODO (justinfagnani): To support nested directives, we'd need to
       // resolve the directive result's values. We may want to offer another
@@ -816,16 +878,19 @@ export class AttributePart {
   _setValue(value: unknown): void;
   _setValue(value: Array<unknown>, from: number): void;
   _setValue(value: unknown | Array<unknown>, from?: number) {
-    const strings = this.__strings;
+    const strings = this.strings;
 
     if (strings === undefined) {
       // Single-value binding case
       const v = this.__resolveValue(value, 0);
+      // Only dirty-check primitives and `nothing`:
+      // `(isPrimitive(v) || v === nothing)` limits the clause to primitives and
+      // `nothing`. `v === this._value` is the dirty-check.
       if (
-        !((isPrimitive(v) || v === nothing) && v === this.__value) &&
+        !((isPrimitive(v) || v === nothing) && v === this._value) &&
         v !== noChange
       ) {
-        this.__commitValue((this.__value = v));
+        this.__commitValue((this._value = v));
       }
     } else {
       // Interpolation case
@@ -843,16 +908,16 @@ export class AttributePart {
         v = this.__resolveValue((value as Array<unknown>)[from! + i], i);
         if (v === noChange) {
           // If the user-provided value is `noChange`, use the previous value
-          v = (this.__value as Array<unknown>)[i];
+          v = (this._value as Array<unknown>)[i];
         } else {
           remove = remove || v === nothing;
           change =
             change ||
             !(
               (isPrimitive(v) || v === nothing) &&
-              v === (this.__value as Array<unknown>)[i]
+              v === (this._value as Array<unknown>)[i]
             );
-          (this.__value as Array<unknown>)[i] = v;
+          (this._value as Array<unknown>)[i] = v;
         }
         attributeValue +=
           (typeof v === 'string' ? v : String(v)) + strings[i + 1];
@@ -877,12 +942,16 @@ export class AttributePart {
 }
 
 export class PropertyPart extends AttributePart {
+  readonly type = PROPERTY_PART;
+
   __commitValue(value: unknown) {
     (this.element as any)[this.name] = value === nothing ? undefined : value;
   }
 }
 
 export class BooleanAttributePart extends AttributePart {
+  readonly type = BOOLEAN_ATTRIBUTE_PART;
+
   __commitValue(value: unknown) {
     if (value && value !== nothing) {
       this.element.setAttribute(this.name, '');
@@ -907,6 +976,7 @@ type EventListenerWithOptions = EventListenerOrEventListenerObject &
  * to add and remove the part as a listener when the event options change.
  */
 export class EventPart extends AttributePart {
+  readonly type = EVENT_PART;
   __eventContext?: unknown;
 
   constructor(...args: ConstructorParameters<typeof AttributePart>) {
@@ -916,7 +986,7 @@ export class EventPart extends AttributePart {
 
   _setValue(newListener: unknown) {
     newListener ??= nothing;
-    const oldListener = this.__value;
+    const oldListener = this._value;
 
     // If the new value is nothing or any options change we have to remove the
     // part as a listener.
@@ -952,16 +1022,16 @@ export class EventPart extends AttributePart {
         newListener as EventListenerWithOptions
       );
     }
-    this.__value = newListener;
+    this._value = newListener;
   }
 
   handleEvent(event: Event) {
-    if (typeof this.__value === 'function') {
+    if (typeof this._value === 'function') {
       // TODO (justinfagnani): do we need to default to this.__element?
       // It'll always be the same as `e.currentTarget`.
-      this.__value.call(this.__eventContext ?? this.element, event);
+      this._value.call(this.__eventContext ?? this.element, event);
     } else {
-      (this.__value as EventListenerObject).handleEvent(event);
+      (this._value as EventListenerObject).handleEvent(event);
     }
   }
 }